+++ bom/pom.xml
@@ -985,6 +985,11 @@
             </dependency>
             <dependency>
                 <groupId>io.helidon.microprofile.testing</groupId>
+                <artifactId>helidon-microprofile-testing-validator</artifactId>
+                <version>${helidon.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.helidon.microprofile.testing</groupId>
                 <artifactId>helidon-microprofile-testing-jaxrs</artifactId>
                 <version>${helidon.version}</version>
             </dependency>
