+++ microprofile/testing/junit5/pom.xml
@@ -65,6 +65,10 @@
             <groupId>io.helidon.microprofile.testing</groupId>
             <artifactId>helidon-microprofile-testing-common</artifactId>
         </dependency>
+        <dependency>
+            <groupId>io.helidon.microprofile.testing</groupId>
+            <artifactId>helidon-microprofile-testing-validator</artifactId>
+        </dependency>
     </dependencies>
 
 </project>
