+++ microprofile/testing/junit5/src/main/java/io/helidon/microprofile/testing/junit5/HelidonJunitExtension.java
@@ -33,7 +33,6 @@ import java.util.Set;
 
 import io.helidon.config.mp.MpConfigSources;
 import io.helidon.config.yaml.mp.YamlMpConfigSource;
-import io.helidon.microprofile.testing.common.JaxRsValidator;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.context.Dependent;
@@ -139,7 +138,8 @@ class HelidonJunitExtension implements BeforeAllCallback,
         }
         validatePerClass();
 
-        JaxRsValidator.validate(testClass);
+        JunitJaxRsValidator junitJaxRsValidator = new JunitJaxRsValidator();
+        junitJaxRsValidator.validate(testClass);
 
         configure(classLevelConfigMeta);
