+++ microprofile/testing/testng/src/main/java/io/helidon/microprofile/testing/testng/TestNgJaxRsValidator.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2023 Oracle and/or its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.helidon.microprofile.testing.testng;
+
+import io.helidon.microprofile.testing.validator.JaxRsValidator;
+
+/**
+ * An implementation for JaxRs Validator for Junit.
+ */
+public class TestNgJaxRsValidator extends JaxRsValidator {
+
+    @Override
+    public void validate(Class<?> testClass) throws RuntimeException {
+        if (checkAddJaxRs(testClass)){
+            DisableDiscovery disableDiscovery = testClass.getAnnotation(DisableDiscovery.class);
+            if (disableDiscovery == null) {
+                throw new RuntimeException("@AddJaxRs annotation may only be used with @DisableDiscovery annotation");
+            }
+        }
+    }
+}
