+++ microprofile/testing/validator/src/main/java/module-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2020, 2023 Oracle and/or its affiliates.
+ * Copyright (c) 2023 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,21 +15,10 @@
  */
 
 /**
- * JUnit5 extension module to run CDI tests.
+ * Tests validators.
  */
-module io.helidon.microprofile.testing.junit5 {
+module io.helidon.microprofile.testing.validator {
 
-    requires io.helidon.config.mp;
-    requires io.helidon.config.yaml.mp;
-    requires io.helidon.microprofile.cdi;
-    requires io.helidon.microprofile.testing.common;
-    requires io.helidon.microprofile.testing.validator;
-    requires jakarta.inject;
-    requires org.junit.jupiter.api;
-
-    requires transitive jakarta.cdi;
-    requires transitive jakarta.ws.rs;
-
-    exports io.helidon.microprofile.testing.junit5;
+    exports io.helidon.microprofile.testing.validator;
 
 }
